requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cache==0.3
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
